test(compiler): use moved utils code, merge tests
