+++ packages/compiler/src/utils/constants.ts
