+++ packages/compiler/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './regex-filters';
 export * from './component-resolvers';
 export * from './load-compiler-cli';
+export { maxWorkers } from './utils';
