--- packages/rsbuild-plugin-angular/src/lib/utils/utils.d.ts
@@ -1,1 +1,0 @@
-export declare const maxWorkers: number;
